Allow SQL Workspace to load
I just wanted to demonstrate the root cause of the issue and share what I did locally to get around the issue and continue playing.

I tried a few different version of sql-wasm and none of them worked so I just copied the one that gets installed via npm. I'm guessing there is a more elegant way to fix this. Feel free to close this PR once you see it.